# UniFlow Academic Service API.openapi+json.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "UniFlow Academic Service API",
        "description": "RESTful API for managing academic data including students, periods, and subjects.\nThis is the core microservice that handles academic entities.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "https://uniflowapi.azure-api.net/academic/v1"
        }
    ],
    "paths": {
...
        "/periods": {
            "get": {
                "summary": "/periods - GET",
                "description": "Get all academic periods for the authenticated user",
                "operationId": "68bc55f7c840211bc0f436e3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": {
                            "enum": [
                                1
                            ],
                            "type": "int",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "enum": [
                                10
                            ],
                            "type": "int",
                            "default": 10
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by period type: first-semester, second-semester, summer, special",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter by year",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "description": "Filter by active status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Periods_With_Pagination"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": "period-2025-2",
                                            "name": "IV Semestre",
                                            "type": "second-semester",
                                            "year": 2025,
                                            "startDate": "2025-08-04T00:00:00.0000000+00:00",
                                            "endDate": "2025-12-05T23:59:59.0000000+00:00",
                                            "studentId": "550e8400-e29b-41d4-a716-446655440000",
                                            "isActive": true,
                                            "createdAt": "2025-08-04T00:00:00.0000000+00:00",
                                            "updatedAt": "2025-08-04T00:00:00.0000000+00:00"
                                        },
                                        {
                                            "id": "period-2025-1",
                                            "name": "III Semestre",
                                            "type": "first-semester",
                                            "year": 2025,
                                            "startDate": "2025-02-03T00:00:00.0000000+00:00",
                                            "endDate": "2025-06-20T23:59:59.0000000+00:00",
                                            "studentId": "550e8400-e29b-41d4-a716-446655440000",
                                            "isActive": false,
                                            "createdAt": "2025-01-20T08:00:00.0000000+00:00",
                                            "updatedAt": "2025-02-01T09:15:00.0000000+00:00"
                                        },
                                        {
                                            "id": "period-2024-verano",
                                            "name": "Verano 2024",
                                            "type": "second-semester",
                                            "year": 2024,
                                            "startDate": "2024-12-15T00:00:00.0000000+00:00",
                                            "endDate": "2025-01-15T23:59:59.0000000+00:00",
                                            "studentId": "550e8400-e29b-41d4-a716-446655440000",
                                            "isActive": false,
                                            "createdAt": "2024-12-15T00:00:00.0000000+00:00",
                                            "updatedAt": "2024-12-15T00:00:00.0000000+00:00"
                                        },
                                        {
                                            "id": "period-2024-2",
                                            "name": "II Semestre",
                                            "type": "second-semester",
                                            "year": 2024,
                                            "startDate": "2024-07-15T00:00:00.0000000+00:00",
                                            "endDate": "2024-11-30T23:59:59.0000000+00:00",
                                            "studentId": "550e8400-e29b-41d4-a716-446655440000",
                                            "isActive": false,
                                            "createdAt": "2024-06-10T10:00:00.0000000+00:00",
                                            "updatedAt": "2024-07-01T12:00:00.0000000+00:00"
                                        }
                                    ],
                                    "pagination": {
                                        "page": 1,
                                        "limit": 10,
                                        "total": 4,
                                        "totalPages": 1,
                                        "hasNext": false,
                                        "hasPrevious": false
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "/periods - POST",
                "description": "Create a new academic period",
                "operationId": "68bc55f7c840211bc0f436e4",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Create_period_response"
                                },
                                "example": {
                                    "id": "period-2025-2",
                                    "name": "II Semestre",
                                    "type": "second-semester",
                                    "year": 2025,
                                    "startDate": "2025-07-14T00:00:00.0000000+00:00",
                                    "endDate": "2025-11-28T23:59:59.0000000+00:00",
                                    "studentId": "550e8400-e29b-41d4-a716-446655440000",
                                    "isActive": false,
                                    "createdAt": "2025-09-06T15:30:00.0000000+00:00",
                                    "updatedAt": "2025-09-06T15:30:00.0000000+00:00"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid period data"
                    }
                }
            }
        },
        "/periods/current": {
            "get": {
                "summary": "/periods/current - GET",
                "description": "Get the current active period for the user",
                "operationId": "68bc55f7c840211bc0f436e5",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/One_Current_Period_Active"
                                },
                                "example": {
                                    "id": "period-2025-1",
                                    "name": "I Semestre",
                                    "type": "first-semester",
                                    "year": 2025,
                                    "startDate": "2025-02-03T00:00:00.0000000+00:00",
                                    "endDate": "2025-06-20T23:59:59.0000000+00:00",
                                    "isActive": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active period found"
                    }
                }
            }
        },
        "/periods/{periodId}": {
            "get": {
                "summary": "/periods/{periodId} - GET",
                "description": "Get specific period details",
                "operationId": "68bc55f7c840211bc0f436e6",
                "parameters": [
                    {
                        "name": "periodId",
                        "in": "path",
                        "description": "Unique period identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Period_By_Id"
                                },
                                "example": {
                                    "id": "period-2025-2",
                                    "name": "IV Semestre",
                                    "type": "second-semester",
                                    "year": 2025,
                                    "startDate": "2025-08-04T00:00:00.0000000+00:00",
                                    "endDate": "2025-12-05T23:59:59.0000000+00:00",
                                    "studentId": "550e8400-e29b-41d4-a716-446655440000",
                                    "isActive": true,
                                    "createdAt": "2025-08-04T00:00:00.0000000+00:00",
                                    "updatedAt": "2025-08-04T00:00:00.0000000+00:00"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Period not found"
                    }
                }
            },
            "put": {
                "summary": "/periods/{periodId} - PUT",
                "description": "Update period information",
                "operationId": "68bc55f7c840211bc0f436e7",
                "parameters": [
                    {
                        "name": "periodId",
                        "in": "path",
                        "description": "Unique period identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Putted_Period"
                                },
                                "example": {
                                    "id": "period-2025-1",
                                    "name": "I Semestre",
                                    "type": "first-semester",
                                    "year": 2025,
                                    "startDate": "2025-02-03T00:00:00.0000000+00:00",
                                    "endDate": "2025-06-20T23:59:59.0000000+00:00",
                                    "studentId": "550e8400-e29b-41d4-a716-446655440000",
                                    "isActive": true,
                                    "createdAt": "2025-01-20T08:00:00.0000000+00:00",
                                    "updatedAt": "2025-02-01T09:15:00.0000000+00:00"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "/periods/{periodId} - DELETE",
                "description": "Delete academic period",
                "operationId": "68bc55f7c840211bc0f436e8",
                "parameters": [
                    {
                        "name": "periodId",
                        "in": "path",
                        "description": "Unique period identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Period deleted successfully"
                    },
                    "409": {
                        "description": "Cannot delete period with associated subjects"
                    }
                }
            }
        },
        "/periods/{periodId}/activate": {
            "patch": {
                "summary": "/periods/{periodId}/activate - PATCH",
                "description": "Activate this period (deactivates others)",
                "operationId": "68bc55f7c840211bc0f436e9",
                "parameters": [
                    {
                        "name": "periodId",
                        "in": "path",
                        "description": "Unique period identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/stats/periods": {
            "get": {
                "summary": "/stats/periods - GET",
                "description": "Get period statistics for the user",
                "operationId": "68bc55f7c840211bc0f436f1",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Period_statistics"
                                },
                                "example": {
                                    "total": 5,
                                    "active": 1,
                                    "current": 1,
                                    "upcoming": 1,
                                    "finished": 3,
                                    "byType": {
                                        "first-semester": 2,
                                        "second-semester": 2,
                                        "summer": 1,
                                        "special": 0
                                    },
                                    "averageDuration": 18
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Periods_With_Pagination": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "required": [
                                "id",
                                "name",
                                "type",
                                "year",
                                "startDate",
                                "endDate",
                                "studentId",
                                "isActive",
                                "createdAt",
                                "updatedAt"
                            ],
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                },
                                "year": {
                                    "type": "integer"
                                },
                                "startDate": {
                                    "type": "string"
                                },
                                "endDate": {
                                    "type": "string"
                                },
                                "studentId": {
                                    "type": "string"
                                },
                                "isActive": {
                                    "type": "boolean"
                                },
                                "createdAt": {
                                    "type": "string"
                                },
                                "updatedAt": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "pagination": {
                        "type": "object",
                        "properties": {
                            "page": {
                                "type": "integer"
                            },
                            "limit": {
                                "type": "integer"
                            },
                            "total": {
                                "type": "integer"
                            },
                            "totalPages": {
                                "type": "integer"
                            },
                            "hasNext": {
                                "type": "boolean"
                            },
                            "hasPrevious": {
                                "type": "boolean"
                            }
                        }
                    }
                },
                "x-apim-schema-name": "Periods With Pagination"
            },
...
        },
        "securitySchemes": {
            "apiKeyHeader": {
                "type": "apiKey",
                "name": "Ocp-Apim-Subscription-Key",
                "in": "header"
            },
            "apiKeyQuery": {
                "type": "apiKey",
                "name": "subscription-key",
                "in": "query"
            }
        }
    },
    "security": [
        {
            "apiKeyHeader": []
        },
        {
            "apiKeyQuery": []
        }
    ]
}